+++ fathom-test/Cargo.toml
@@ -10,7 +10,7 @@ license = "Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dev-dependencies]
+[dependencies]
 assert_fs = "1.0"
 codespan-reporting =  "0.11"
 fathom = { version = "0.1.0", path = "../fathom" }
